cavm: support GPIO interrupts in application
This patch ports the GPIO interrupts to application feature to
OcteonTx2 family. As part of the port lot of cleanup is done to fit into
current design of ATF.

Brief feature information:
User space program registers a handler for a specific
GPIO and core number. Four SPI numbers are reserved for 4 cores. On a
GPIO interrupt the core takes an exception in EL3, checks for the GPIO
line interrupt state and calls the registered user space handler. After
user space handler the control returns to EL3 by taking exception
executing WFI instruction in EL0. This will complete the interrupt
handling.

Change-Id: Id00360eb87c2938ca459026601a9a314ee38a482
Signed-off-by: Radha Mohan Chintakuntla <rchintakuntla@cavium.com>
Signed-off-by: Alex Belits <abelits@cavium.com>
Signed-off-by: Tirumalesh Chalamarla <tchalamarla@cavium.com>