drivers: marvell: Move plat/marvell/octeontx/otx2/cavm_rvu.c to drivers
cavm_rvu.c is moved to drivers/marvel. There is not additional
subdirectory for OcteonTX2, because this way sources are organized
by Marvell in mainline ATF repository. In cavm_rvu.c is additional
comment at the begining of file, that this driver is for OcteonTX2.

octeontx2_clear_lf_to_pf_mapping() and octeontx_rvu_init() are moved
from cavm_common.h to cavm_rvu.h, because this is the file which
should contains API for RVU.

Change-Id: I06aad5880393a49f09f63e3a9585fd78763fa9f2
Signed-off-by: Konrad Adamczyk <konrad.adamczyk@cavium.com>