octeontx2: flr: Add support for load/store register pair instructions
This patch add support for load and store rgister pair instructions.
SIMD instructions are excluded, because it will be hard to implement
and probably no one ever will use registers trapped by FLR in this
kind of instructions. Except that every variant of pair
store and load instruction is implemented.

Change-Id: I913f4de04e033b633e23c1ae3bbbe7a82832eb4d
Signed-off-by: Konrad Adamczyk <kadamczyk@marvell.com>