plat: octeontx2: parser: Update the QLM indexing to start parsing from QLM 0
Right now, parser starts looking for ethernet mode from MAX GSERN lane(in case of T96, it starts at 7). This was done to take care that if both QLM 3 and QLM 7 are configured as ethernet mode, only 1 can be configured to CGX0 and since BDK configures QLM 7 to CGX 0 in this case, parser is implemented to reverse parse.

in case of FIO board, this conflict will not occur. In case of EBB9604, MCU doesnt allow QLM 3 and QLM 7 to be configured to ethernet mode at the same time.

With this patch, still there is a possibility that on ASIM, user can configure from BDK DT to have both QLM 3 and QLM 7 to be of ethernet mode unknowingly and in that case, QLM3 will take high priority and be mapped to CGX 0 (unlike BDK)

Change-Id: I6562e50c1ef6a217bda4c4c1b98b49bb696f2a5a
Signed-off-by: Nithya Mani <nithya.mani@cavium.com>
Reviewed-on: https://sj1git1.cavium.com/1453
Tested-by: Chandrakala Chavva <Chandrakala.Chavva@cavium.com>
Reviewed-by: Chandrakala Chavva <Chandrakala.Chavva@cavium.com>