drivers: marvell: cgx: Update MODE_CHANGE command to use mode bitmask
With speed change feature support in the previous version,
only the speed and autoneg parameter could be changed. This
patch adds support for the user to choose the MODE as
different MODES/protocols are applicable for each of the
speed. Hence add MODE parameter to the command args.

For each mode, ethtool provides standard BITMASK and
mapping is provided for each mode supported.

Change-Id: I9c4fc6004677d97962a8711d4939505306f247ec
Signed-off-by: Nithya Mani <nithya.mani@cavium.com>
Reviewed-on: https://sj1git1.cavium.com/15600
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Tested-by: Chandrakala Chavva <cchavva@marvell.com>