mrvl: Handle exception when WFI is executed by EL0 ISR
The commit 2887fa created an issue where EL0 interrupt handling was
causing unhandled exception because the exception caused when WFI is
executed by EL0 was not being handled. The el3_exit_wfi needs to be
called to handle this situation for a clean return from EL3 to mark end
of interrupt handling.

Change-Id: Ia95c06b9c50bc5c2f417fb5215cd89d692bb935e
Signed-off-by: Radha Mohan Chintakuntla <radhac@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/22020
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Nithya Mani <Nithya.Mani@cavium.com>
Reviewed-by: Tomasz Michalec <tomasz@marvell.com>
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>