drivers: marvell: qlm: Move ATF functions to qlm_gserx_ops
Functions that are not in BDK qlm_gserc.c and qlm_gserr.c files are
moved to qlm_gserc_ops.c and qlm_gserr_ops.c. They are moved because
qlm_gserc.c and qlm_gserr.c files are synchronized with BDK and cannot
be modified manually (this will break update process).

Change-Id: I3f94b60287ee19d1a2e2c971b7b9d83420cf2994
Signed-off-by: Tomasz Michalec <tomasz@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/24515
Reviewed-by: Nithya Mani <Nithya.Mani@cavium.com>
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Tested-by: Chandrakala Chavva <cchavva@marvell.com>