EHF: add workaround for anomaly where GIC returns IDLE priority
Occasionally, the GIC will return the 'IDLE' priority
for an interrupt which is in-service; this prevents
EHF from locating the correct exception handler,
resulting in a panic.
In this case, EHF will now locate the correct priority
from the interrupt array, thus allowing the correct
exception handler to be located and invoked.

Change-Id: I34dcc970557b8cf59aacacf47e660a4dc2acfe1d
Signed-off-by: Rick Farrington <Ricardo.Farrington@cavium.com>
Reviewed-on: https://sj1git1.cavium.com/24601
Tested-by: Chandrakala Chavva <cchavva@marvell.com>
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>