drivers: marvell: gti: Fix for -fno-common
In newer GCC (10 and above), -fno-common is default,
this means there are no more common symbols by default.
This means tentative definition become normal definition
at the end of the translation unit.  And if there are
more than one definition, the link will fail with multiple
symbols defined.  In this case del3t_trap_handler
was a tentative definition in gti_watchdog.h; we need
to turn this into an external declaration instead.

Change-Id: I92a1ed1cf6a43635b082faf9580cc6f0af72c28a
Signed-off-by: Andrew Pinski <apinski@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/25412
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>