drivers: marvell: cgx: Add CPRI mode change command
This patch increases FW version to 1.1 for the support
of new command. Kernel/U-boot FW version also needs
to be updated.

* Also, added check in CGX MODE change command to not
process CPRI mode change request

* The command expects GSERC and lane as index with the
requested baud rate. This command handles from change
from ethernet to CPRI and different baud rates in CPRI
and not from CPRI to CGX.
* If change request to baud rate for the lane that is not
in the same set of baud rate group, it will return error.
* If change request to baud rate for the lane is changed
from ethernet to CPRI, other lane in DLM is also changed.

FIXME: cgx.md to be updated with this command usage.

Change-Id: I199760bb6bfacba58fbb23b6a1a066f6f098fec0
Signed-off-by: Nithya Mani <nithya.mani@cavium.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/26414
Tested-by: Chandrakala Chavva <cchavva@marvell.com>
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>