context: Set the SCTLR_ELx.nTWI bit before jump to EL0
Newer kernels are handling WFI executed by EL0 but we need to handle the
WFI exception here in EL3 so that the user space interrupt can return to
EL3 to finish the ISR and get out of interrupt context. This patch
temporarily disables kernel handling of WFI exception by EL0.

Signed-off-by: Radha Mohan Chintakuntla <radhac@marvell.com>
Change-Id: I50416c11e2b11f415e64ce945b0e503b11372e84
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/27355
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>