drivers: marvell: cgx: update lane index for CPRI mode change
When changing from E-CPRI to CPRI, lane index needs to be
passed as -1 so both lanes of a DLM are changed to CPRI
mode. If respective lane index is passed to set_mode API,
and if other mode is in E-CPRI mode, there are issues
with mode change.

Also, adjust calculation of lane index as the lane
mapping for DLM combinations changed recently.

Change-Id: Icdde272771ec04454cff06aa0c3f9dabd12ab422
Signed-off-by: Nithya Mani <nithya.mani@cavium.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/27371
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>