libfdt: Fix alignment exceptions when retrieving fdt64_t values.
The current method of retrieving an fdt64_t requires a value
to be converted; this requires dereferencing the FDT property
pointer (to obtain the value needed for conversion).
Such a dereference can result in an alignment exception.
Add support for retrieving an fdt64_t directly from an FDT
property pointer.

Change-Id: I9e42f56752e52cc4779aef51cbf1cae81dffd2b3
Signed-off-by: Rick Farrington <rfarrington@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/27375
Tested-by: Chandrakala Chavva <cchavva@marvell.com>
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>