Marvell: Add Ethernet fast-link and activity LEDs
When GPIO LEDs configured as Ethernet LEDs in BDK, added support
to make the link LEDs glow and Activity LED link according to
the link up and activity conditions.
With this changes ATF will have the GPIO LED Ethernet
configurations picked up from device tree file in
octeontx2_fill_cgx_details.
GPIO connected Activity LEDs are driven by a fixed 4 Hz frequency
from the GPIO clock generator, and if there is activity within
the last second on the Tx or Rx, the GPIO will be enabled, and
disabled if there is none.

Change-Id: Ia0ed333b976f60e520edfe10dd0ba1f33763d65b
Signed-off-by: svenkatachalam <Selvam.Venkatachalam@cavium.com>
Signed-off-by: Felix Manlunas <fmanlunas@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/26270
Reviewed-by: Nithya Mani <Nithya.Mani@cavium.com>
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>