plat: marvell: otx2: add support for BERT in NS preserved memory
The Boot Error Record Table (BERT) allows fatal RAS errors to be
reported to Linux on the next boot.  This change adjusts the
Non-Secure Preserved Memory region to accommodate both the
'ramoops' data as well as the BERT.  The size of the BERT
memory is controlled via a Device Tree 'reserved-memory'
entry.

Change-Id: I67690c832fe2e86ff91ecec77485d78f0865093b
Signed-off-by: Rick Farrington <rfarrington@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/31405
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>