drivers: marvell: rvu: fix provisioning problem
Provisioning of "available" RVU PFs could overwrite [legacy]
"fixed" RVU PFs.

The RVU provisioning overwrote the "fixed" SSO PF entry, resulting
in no SSO VFs available in Linux (see on a Loki platform).

After the "fixed" and network RVU PFs are provisioned, the remaining
RVU PFs are provisioned to SSO & NPA in a 3:1 ratio.
The set of "remaining" RVU PFs consists of all unused entries reserved
for ETH usage as well as any unused "fixed" reservations.  This pool
of "remaining" RVU PFs is not sequentially ordered.  However, the code
was treating it as such, resulting in an RVU provisioning overwrite.

This regression was introduced in the following commit:
  2ea3ef58: Integration: plat: marvell: rvu: fix provisioning when no
            net devs are present

Change-Id: Ie28a081769a2af92995dac38657d450068c473ed
Signed-off-by: Rick Farrington <rfarrington@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/39734
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>