plat: marvell: otx2: support relocation of GHES area to top-of-ram
Optionally relocate the GHES base addresses to the top of OS DRAM so
as not to fragment the system memory.  The FDT 'ghes-hest' reg
property determines if relocation occurs; if this value is non-zero,
then the specified address is used (legacy).  Otherwise the
relocation is performed.

Also, invoke all functions which adjust the FDT from new function
'plat_adjust_fdt()'.

Change-Id: I3cd1c94b0c09a8d87a03a228c2d3e9ff2e3f51a0
Signed-off-by: Rick Farrington <rfarrington@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/42922
Tested-by: Chandrakala Chavva <cchavva@marvell.com>
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>