plat: marvell: add SMC call to list BPHY irqs
This patch provides information about supported
interrupts for given platform available to the
non-secure world.
Available interrupt is represented by a bit set of
the returned value on a position indicated by
interrupt number. Maximum supported number of
interrupts is limited by bits available in a single
register, that is 64. This limit can be extended
if necessary, however, current range seems enough
for some time.

This patch enables this feature on the following
platforms: loki, f95, f95mm, f95o.

Signed-off-by: Jakub Palider <jpalider@marvell.com>
Change-Id: I15a056d95b0ad28b2b94a403925280a6277847e0
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/47723
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Tested-by: Chandrakala Chavva <cchavva@marvell.com>