plat marvell: optee: Add workaround to avoid lockup in system reset path
Added the workaround to avoid lockup in TF-A when
calling system reset hook functions of optee.

For the command "echo c >/proc/sysrq-trigger"
hangs in opteed_enter_sp(), its the watchdog that
triggers the reboot and not the system reset.
This is a workaround until we find a fix.

Change-Id: I44a5fc3a6d79cdb07f4ae5ee29eae16f9334c164
Signed-off-by: Anil Kumar Reddy <areddy3@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/65156
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Tested-by: Chandrakala Chavva <cchavva@marvell.com>
(cherry picked from commit 93bb3adf69d402c9bbf038c1616184486d7a5284)
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/65228
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>